Refactor Node-RED S7 Suite and enhance functionality
- Updated .gitignore to include development artifacts and logs.
- Modified docker-compose.yml to use a specific Node-RED image and adjusted entrypoint for improved package management.
- Removed Dockerfile as the build process is now handled directly in docker-compose.
- Bumped package version to 0.2.0 and updated author information in package.json.
- Enhanced README with new features, including the addition of the s7-control node for PLC control actions.
- Improved example flows to demonstrate new output modes and added support for reading buffers, structs, and bits.
- Refined backend interfaces and added PLC control methods in the snap7 backend.
- Updated address parser to support additional S7 data types and improved error handling.
- Enhanced connection manager with request timeout handling and improved state management.
- Updated icons for better visual representation in the Node-RED palette.

+++ test/unit/core/address-parser-extra.test.ts
@@ -12,13 +12,13 @@ describe('address-parser - extra coverage', () => {
   it('parses C area', () => {
     const addr = parseAddress('C0');
     expect(addr.area).toBe('C');
-    expect(addr.dataType).toBe('BYTE');
+    expect(addr.dataType).toBe('WORD');
   });
 
   it('parses T area', () => {
     const addr = parseAddress('T0');
     expect(addr.area).toBe('T');
-    expect(addr.dataType).toBe('BYTE');
+    expect(addr.dataType).toBe('WORD');
   });
 
   it('toNodes7Address for INT uses W', () => {
